+++ ChallengePromptMaker.desktop
@@ -1,0 +1,8 @@
+[Desktop Entry]
+Type=Service
+ServiceTypes=Krita/PythonPlugin
+X-KDE-Library=ChallengePromptMaker
+X-Python-2-Compatible=false
+X-Krita-Manual=Manual.html
+Name=Challenge Prompt Maker
+Comment=A docker that generate prompt randomly based off a list.
