+++ ChallengePromptMaker/__pycache__/CPM_CategoryList.cpython-38.pyc
