+++ ChallengePromptMaker/__pycache__/CPM_ColorManager.cpython-38.pyc
