--- category.json
@@ -1,197 +1,0 @@
-{
-    "Fruits": [
-        "Apple",
-        "Mango",
-        "Blueberry",
-        "Papaya",
-        "Apricot",
-        "Banana",
-        "Grapefruit",
-        "Lemon",
-        "Lime",
-        "Pineapple",
-        "Jackfruit",
-        "Melon",
-        "Coconut",
-        "Avocado",
-        "Peach",
-        "Kiwi",
-        "Blackcurrant",
-        "Blackberry",
-        "Cherry",
-        "Fig",
-        "Lychee",
-        "Nectarine",
-        "Passionfruit",
-        "Raspberry ",
-        "Pomegranate",
-        "Mulberry",
-        "Starfruit ",
-        "Guava",
-        "Pomelo",
-        "Cranberry",
-        "Curuba",
-        "Soursop",
-        "Melon Cantaloupe",
-        "Granadilla",
-        "Durian",
-        "Dragon Fruit",
-        "Longan",
-        "Rambutan",
-        "Rose Apple",
-        "Star Apple",
-        "Tamarind",
-        "Persimmon",
-        "Mangosteen",
-        "Pear",
-        "Strawberry"
-    ],
-    "Vegetables": [
-        "Cucumber",
-        "Yellow Onion",
-        "Red Onion",
-        "Garlic",
-        "Carrot",
-        "Red Cabbage",
-        "White Cabbage",
-        "Radish",
-        "Eggplant",
-        "Mushroom",
-        "Artichoke",
-        "Corn",
-        "Broccoli",
-        "Cauliflower",
-        "Celery",
-        "Red Chili",
-        "Green Chili",
-        "Sweet Potato",
-        "Asparagus",
-        "Pumpkin",
-        "Fennel",
-        "Spring Onion",
-        "Turnip",
-        "Lettuce",
-        "Zucchini ",
-        "Brussels Sprout",
-        "Tomato",
-        "Potato",
-        "Pea",
-        "Spinach",
-        "Beetroot ",
-        "Bell Pepper",
-        "Leek",
-        "Ginger",
-        "Squash",
-        "Cilantro",
-        "Napa Cabbage ",
-        "Kale ",
-        "Yam",
-        "Bitter Melon",
-        "Taro",
-        "Kohlrabi",
-        "Beansprouts"
-    ],
-    "Animals": [
-        "Cat",
-        "Dog",
-        "Gold Fish",
-        "Monitor Lizard",
-        "Bearded Dragon",
-        "Lion",
-        "Hyena",
-        "Tiger",
-        "Sugar Glider",
-        "Squirel",
-        "Rat",
-        "Parrot",
-        "Cockatiel",
-        "Leopard",
-        "Cobra",
-        "Boa",
-        "African wild dog"
-    ],
-    "Trees and Shrubs": [
-        "Bamboo",
-        "Oak",
-        "Acacia",
-        "Dragon Tree",
-        "Sakura Tree",
-        "Pine",
-        "Palm",
-        "Maple",
-        "Birch",
-        "Aspen",
-        "Ash Tree",
-        "Rosewood",
-        "Mahogany",
-        "Ebony"
-    ],
-    "Gems": [
-        "Alexandrite",
-        "Amethyst",
-        "Aquamarine",
-        "Citrine",
-        "Diamond",
-        "Emerald",
-        "Garnet",
-        "Jade",
-        "Lapis Lazuli",
-        "Moonstone",
-        "Morganite",
-        "Onyx",
-        "Opal",
-        "Paraiba Tourmaline",
-        "Pearls",
-        "Peridot",
-        "Ruby",
-        "Sapphire",
-        "Spinel",
-        "Tanzanite",
-        "Topaz",
-        "Tourmaline",
-        "Turquoise",
-        "Zircon"
-    ],
-    "Vehicles": [
-        "Car",
-        "Boat",
-        "Cruise",
-        "Spaceship",
-        "Helicopter",
-        "6 Wheeler Truck",
-        "Motorcycle",
-        "Bicycle",
-        "Hoverboard"
-    ],
-    "Bread and Dessert": [
-        "Loaf Bread",
-        "Cake",
-        "Donut",
-        "Eclair",
-        "Smores",
-        "Rye Bread",
-        "Ice Cream",
-        "Froyo"
-    ],
-    "Food": [
-        "Burger",
-        "Cheese Burger",
-        "Pasta",
-        "Spaghetti",
-        "Alfredo",
-        "Pizza",
-        "Fried Chicken",
-        "Ramen",
-        "Pho"
-    ],
-    "Magical Creatures": [
-        "Vampire",
-        "Werewolf",
-        "Fairy",
-        "Will-o-wisp",
-        "Selky",
-        "Yeti",
-        "Mermaid",
-        "Sirene"
-    ]
-}
